+++ src/Plugins/ToolBar.tsx
@@ -70,6 +70,7 @@ import { InsertImageDialog } from './ImagePlugin';
 import { InsertInlineImageDialog } from './InlineImage';
 import { InsertLinkPlugin } from './InsertLink';
 import { INSERT_PAGE_BREAK } from './PageBreak';
+import { PageSetupPlugin } from './PageSetup';
 import { TableItemPlugin } from './Table';
 import { YoutubeUploadPlugin } from './Youtube';
 
@@ -105,7 +106,8 @@ type ToolbarToken =
   | 'FontSize'
   | 'Decorators'
   | 'CodeBlock'
-  | 'Align';
+  | 'Align'
+  | 'PageSetup';
 
 const ALLOWED_TOKENS: Record<ToolbarToken, true> = {
   '|': true,
@@ -124,6 +126,7 @@ const ALLOWED_TOKENS: Record<ToolbarToken, true> = {
   Decorators: true,
   CodeBlock: true,
   Align: true,
+  PageSetup: true,
 };
 
 /**
@@ -677,6 +680,16 @@ export const ToolBarPlugins = (props: IEditorProps) => {
           </Dropdown>
         );
       }
+
+      case 'PageSetup':
+        return (
+          <PageSetupPlugin
+            key={key}
+            disabled={!isEditable || props.readOnly}
+            value={props.pageSetup}
+            onChange={props.onPageSetupChange}
+          />
+        );
 
       default:
         return null;